+++ hygen/_templates/component/help/index.ejs.t
@@ -1,0 +1,9 @@
+---
+message: |
+
+    hygen {bold component new} --name [NAME]
+
+    Generates a component
+
+       --name             The component name in kebab-case ({bold required})
+---