+++ hygen/_templates/layer/new/inject_import_layer.ejs.t
@@ -2,7 +2,7 @@
 inject: true
 to: src/components/layers/index.js
 skip_if: "import <%= h.changeCase.pascalCase(name) %>"
-before: "// Auto import"
+before: // \[hygen\] Import layers
 ---
 <% const comp = h.changeCase.pascalCase(name) -%>
-import <%= comp %> from './<%= comp %>';
+import <%= h.changeCase.pascalCase(name) %> from './<%= comp %>';