+++ hygen/_templates/slice/help/index.ejs.t
@@ -1,0 +1,9 @@
+---
+message: |
+
+    hygen {bold component new} --name [NAME]
+
+    Generates a slice
+
+       --name             The slice name in kebab-case ({bold required})
+---