+++ template-sample-app/template/src/components/views/Kpi.js
@@ -58,7 +58,7 @@ function Kpi() {
     };
   }, [dispatch]);
 
-  // Auto import useEffect
+  // [hygen] Add useEffect
 
   const onTotalRevenueWidgetError = (error) => {
     dispatch(setError(`Error obtaining total revenue: ${error.message}`));