+++ template-sample-app/template/src/components/views/stores/StoresList.js
@@ -21,7 +21,7 @@ function StoresList() {
     dispatch(setBottomSheetOpen(false));
   }, [dispatch]);
 
-  // Auto import useEffect
+  // [hygen] Add useEffect
 
   const onTotalRevenueWidgetError = (error) => {
     dispatch(setError(`Error obtaining total revenue: ${error.message}`));